+++ laktory/resources/quickstart-stacks/workflows/notebooks/dlt_laktory_pl.py
@@ -1,3 +1,4 @@
+# Databricks notebook source
 # COMMAND ----------
 import json